devex: move all web-client persistence methods to web-client folder
