+++ web-client/src/persistence/dynamsoft/getScannerInterface.ts
