+++ web-client/src/persistence/localStorage/removeItem.ts
