+++ web-client/src/persistence/s3/getDocument.ts
