+++ web-client/src/persistence/s3/getPdfFromUrl.test.ts
@@ -1,4 +1,4 @@
-import { applicationContext } from '../../business/test/createTestApplicationContext';
+import { applicationContext } from '../../../../shared/src/business/test/createTestApplicationContext';
 import { getPdfFromUrl } from './getPdfFromUrl';
 
 describe('getPdfFromUrl', () => {
