+++ web-client/src/persistence/s3/uploadDocumentFromClient.test.ts
@@ -1,4 +1,4 @@
-import { applicationContext } from '../../business/test/createTestApplicationContext';
+import { applicationContext } from '../../../../shared/src/business/test/createTestApplicationContext';
 import { uploadDocumentFromClient } from './uploadDocumentFromClient';
 
 describe('uploadDocument', () => {
