+++ web-client/src/persistence/s3/uploadDocumentFromClient.ts
