+++ web-client/src/persistence/s3/uploadPdfFromClient.test.ts
@@ -1,4 +1,4 @@
-import { applicationContext } from '../../business/test/createTestApplicationContext';
+import { applicationContext } from '../../../../shared/src/business/test/createTestApplicationContext';
 import { uploadPdfFromClient } from './uploadPdfFromClient';
 
 describe('uploadPdfFromClient', () => {
